Rename class/tests, move data source to new file
